web/requirements.txt
Django==1.9.5
djangorestframework==3.5.0
django-oauth-toolkit
django-rest-swagger
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
Pillow